backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Max Platonov"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "sqlmodel (>=0.0.24,<0.0.25)",
    "alembic (>=1.15.2,<2.0.0)",
    "uv (>=0.6.12,<0.7.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "redis (>=4.2.0rc1,<5.0.0)",
    "fastapi-cache2[redis] (>=0.2.2,<0.3.0) ; python_version >= \"3.8\" and python_version < \"4.0\"",
    "pydantic[email] (>=2.11.2,<3.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "psycopg-binary (>=3.2.6,<4.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "psycopg (>=3.2.6,<4.0.0)",
    "tenacity (>=9.1.2,<10.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
